修改调整解析雇员服务协议参数



查看合并请求 !95

+++ SalaryGrantSiteService/salary-grant-site-service-api/src/main/java/com/ciicsh/gto/fcbusinesscenter/salarygrant/siteservice/api/dto/PayrollCalcResultDTO.java
@@ -57,7 +57,7 @@ public class PayrollCalcResultDTO extends PagingDTO implements Serializable {
     /**
      * 离职年限
      */
-	private Integer leavingYears;
+	private Double leavingYears;
     /**
      * 实发工资
      */
@@ -199,11 +199,11 @@ public class PayrollCalcResultDTO extends PagingDTO implements Serializable {
 		this.refBatchId = refBatchId;
 	}
 
-	public Integer getLeavingYears() {
+	public Double getLeavingYears() {
 		return leavingYears;
 	}
 
-	public void setLeavingYears(Integer leavingYears) {
+	public void setLeavingYears(Double leavingYears) {
 		this.leavingYears = leavingYears;
 	}
